requirements.txt
flask~=2.3.2
requests~=2.31.0
colorama~=0.4.4
pip~=23.2.1
attrs~=23.1.0
Pillow~=10.0.0
docutils~=0.16
Jinja2~=3.1.2
setuptools~=65.5.0
selenium~=4.10.0
mysql-connector-python~=8.1.0
tqdm~=4.66.1
SQLAlchemy~=2.0.22
alembic~=1.12.1
mysqlclient
